SIG-8: checkboxes, groupboxes, styling, new migrations, accurate field positioning

+++ lib/generators/active_record/templates/doc_tag_migration.rb
@@ -3,10 +3,10 @@ class SignatureCreate<%= table_name.camelize.singularize %>Tags < ActiveRecord::
     create_table(:<%= table_name.singularize %>_tags) do |t|
       t.string :name
       t.string :value
-      t.integer :x
-      t.integer :y
-      t.integer :width
-      t.integer :height
+      t.float :x
+      t.float :y
+      t.float :width
+      t.float :height
       t.string :tag_type
       t.integer :page
       t.boolean :white_bg
